+++ frontend/src/components/settings/AppearancePanel.css
@@ -1,0 +1,71 @@
+/* Settings → Appearance panel.
+   Mirrors the layout of PerformancePanel (Wave 2 INST-12) so all the
+   per-section Settings sub-panels read consistently. */
+
+.appearance-panel {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  padding: 14px 16px;
+  border: 1px solid rgba(255, 255, 255, 0.06);
+  border-radius: 10px;
+  background: rgba(255, 255, 255, 0.02);
+}
+
+.appearance-panel__title {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  margin: 0;
+  font-size: 13px;
+  font-weight: 600;
+  color: rgba(255, 255, 255, 0.85);
+  letter-spacing: 0.02em;
+}
+
+.appearance-panel__row {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  flex-wrap: wrap;
+}
+
+.appearance-panel__label {
+  min-width: 100px;
+  font-size: 12px;
+  color: rgba(255, 255, 255, 0.65);
+}
+
+.appearance-panel__themes {
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.appearance-panel__theme-dot {
+  width: 18px;
+  height: 18px;
+  border-radius: 50%;
+  border: 2px solid rgba(255, 255, 255, 0.18);
+  background: var(--dot-color, #888);
+  cursor: pointer;
+  padding: 0;
+  transition: transform 0.15s ease, border-color 0.15s ease, box-shadow 0.15s ease;
+}
+
+.appearance-panel__theme-dot:hover {
+  transform: scale(1.1);
+  border-color: rgba(255, 255, 255, 0.4);
+}
+
+.appearance-panel__theme-dot.is-active {
+  border-color: rgba(255, 255, 255, 0.9);
+  box-shadow: 0 0 0 2px rgba(255, 255, 255, 0.15);
+}
+
+.appearance-panel__help {
+  margin: 4px 0 0;
+  font-size: 11.5px;
+  line-height: 1.5;
+  color: rgba(255, 255, 255, 0.5);
+}
